requirements.txt
fastapi
uvicorn
requests
scholarly
networkx
node2vec
gensim
scikit-learn
sqlalchemy
python-dotenv
sentence-transformers
aiohttp
